create db connection config and change item model price from float32 to int to match db schema

+++ models/itemsModel.go
@@ -4,7 +4,7 @@ type Items struct {
 	ID string `json:"id"`
 	Image string `json:"image"`
 	Name string `json:"name"`
-	Price float32 `json:"price"`
+	Price int `json:"price"`
 	Stock int `json:"stock"`
 	Category string `json:"category"`
 }
